part.scale change from double to in
git-svn-id: http://eharmony-photofeature.googlecode.com/svn/trunk@11 0b1a5a23-84ec-7007-e15f-aefa3a085764

+++ eHfacemodel.h
@@ -33,7 +33,7 @@ typedef struct part {
 	/*cache for later use, not from file*/
 	int sizy;
 	int sizx;
-	double scale;
+	int scale;
 	int startx;
 	int starty;
 	int step;
